Instantiation of point lights

+++ src/pointLight.cpp
@@ -19,9 +19,9 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
     exponent = exp; 
 }
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColorLocation, 
-                GLfloat diffuseIntensityLocation, GLfloat positionLocation,
-                GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColorLocation, 
+                GLuint diffuseIntensityLocation, GLuint positionLocation,
+                GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
     glUniform3f(ambientColorLocation, color.x, color.y, color.z); // in the shader we will have a vec3 and it will be bound to ambientColorLocation
     glUniform1f(ambientIntensityLocation, ambientIntensity);
